# frontend/src/types/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "个人化资讯监控管理系统 API",
    "title": "Personal Information Monitor",
    "version": "1.4.3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/system/doctor": {
      "get": {
        "description": "Perform a full system diagnostic audit.",
        "operationId": "get_system_doctor_api_system_doctor_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get System Doctor Api System Doctor Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get System Doctor",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/features": {
      "get": {
        "description": "Expose static + env-driven feature flags for clients.",
        "operationId": "get_runtime_features_api_system_features_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Runtime Features Api System Features Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Runtime Features",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/system/queue": {
      "get": {
        "description": "Get task and fetch status for observability.",
        "operationId": "get_queue_status_api_system_queue_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Queue Status Api System Queue Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Queue Status",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/score-vocab/reload": {
      "post": {
        "description": "Reload YAML-backed score vocabulary without restarting PIM.",
        "operationId": "reload_score_vocab_runtime_api_system_score_vocab_reload_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reload Score Vocab Runtime Api System Score Vocab Reload Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Reload Score Vocab Runtime",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/search/rebuild": {
      "post": {
        "description": "Manually trigger a full rebuild of the search index.",
        "operationId": "rebuild_search_index_api_system_search_rebuild_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Rebuild Search Index Api System Search Rebuild Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Rebuild Search Index",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/source-types": {
      "get": {
        "description": "Return the canonical source-type catalog.\n\nSingle source of truth for UI dropdowns / tabs / prompt lists — clients\nshould call this at startup and reuse the bundled copy as a fallback.",
        "operationId": "get_source_types_api_system_source_types_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_disabled",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Disabled",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Source Types Api System Source Types Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Source Types",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/support-bundle": {
      "get": {
        "description": "Generate a redacted diagnostic bundle for manual sharing.",
        "operationId": "download_support_bundle_api_system_support_bundle_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Download Support Bundle",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/update-check": {
      "get": {
        "description": "Check GitHub releases and report whether a newer PIM version exists.",
        "operationId": "get_update_check_api_system_update_check_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_prerelease",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Prerelease",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Update Check Api System Update Check Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Update Check",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/upgrade": {
      "get": {
        "description": "Return the latest UI-triggered upgrade status and log tail.",
        "operationId": "get_upgrade_status_api_system_upgrade_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Upgrade Status Api System Upgrade Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Upgrade Status",
        "tags": [
          "system"
        ]
      },
      "post": {
        "description": "Start a detached ``./pim upgrade`` runner if one is not already active.",
        "operationId": "start_upgrade_api_system_upgrade_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Start Upgrade Api System Upgrade Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Start Upgrade",
        "tags": [
          "system"
        ]
      }
    },
...
  }
}